requirements.txt
Django==5.2.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
psycopg2-binary
python-dotenv==1.0.0
Pillow==10.4.0
